Requested changes from PR
