+++ src/layouts/AppContainer.tsx
@@ -12,7 +12,8 @@ import Header from "layouts/Header";
 import SidebarMenu from "layouts/SidebarMenu";
 import { IRootState } from "reducers";
 import { dismissNotification, INotificationsState, NotificationStatus, showNotification, INotification } from "reducers/notifications";
-import { getCachedAccount, cacheWeb3Info, logout, pollForAccountChanges, pollSubgraphUpdating } from "arc";
+import { getCachedAccount, cacheWeb3Info, logout, pollForAccountChanges } from "arc";
+import { pollSubgraphUpdating } from "lib/subgraphHelpers";
 import ErrorUncaught from "components/Errors/ErrorUncaught";
 import { parse } from "query-string";
 import * as React from "react";
